Move user interface out of game.py
