+++ aesthetic_ants/penguin/game.py
@@ -2,12 +2,12 @@ import pyglet
 
 from .constants import CollisionType
 from .enemy import BigEnemy, FastEnemy
-from .object import Object
 from .player import Player
 from .resources import LEVEL_1
 from .snowball import Snowball
 from .space import Space
 from .tile_layer import TileLayer
+from .ui import GameOverScreen, ScoreLabel
 from .utils import keys
 
 
@@ -124,33 +124,3 @@ class Game(pyglet.window.Window):
             self.score_label.label.text = str(self.score)
 
         enemy.on_collision_snowball(snowball)
-
-
-class GameOverScreen(Object):
-    def __init__(self, window):
-        self.label = pyglet.text.Label("Game Over!",
-                                       font_name="Times New Roman",
-                                       font_size=36,
-                                       x=window.width // 2,
-                                       y=window.height // 2,
-                                       anchor_x='center',
-                                       anchor_y='center')
-
-    def add_to_space(self, space):
-        super().add_to_space(space)
-        self.label.batch = space.batch
-
-
-class ScoreLabel(Object):
-    def __init__(self, window):
-        self.label = pyglet.text.Label("0",
-                                       font_name="Times New Roman",
-                                       font_size=18,
-                                       x=10,
-                                       y=window.height - 10,
-                                       anchor_x='left',
-                                       anchor_y='top')
-
-    def add_to_space(self, space):
-        super().add_to_space(space)
-        self.label.batch = space.batch
